requirements.txt
tensorflow==2.13
pandas
Pillow
opencv-python
flask
